apps/api/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ysdevidx-api"
version = "0.1.0"
description = "Yonsei Dev Index API"
requires-python = ">=3.11"
dependencies = [
  "alembic>=1.13.2",
  "fastapi>=0.115.0",
  "httpx>=0.27.2",
  "psycopg[binary]>=3.2.3",
  "pydantic-settings>=2.4.0",
  "python-dotenv>=1.0.1",
  "sqlalchemy>=2.0.35",
  "uvicorn[standard]>=0.30.6"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.3",
  "ruff>=0.6.9"
]

[tool.setuptools]
packages = ["app"]
